+++ backend/src/orderitem/orderitem.service.ts
@@ -1,20 +1,30 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { createOrderItemDto } from './dto/create-orderitem.dto';
+import { CreateOrderItemDto } from './dto/create-orderitem.dto';
 
 @Injectable()
 export class OrderitemService {
   constructor(private readonly prisma: PrismaService) {}
-  async createOrderItem(data: createOrderItemDto) {
-    const orderItem = await this.prisma.orderItem.create({ data });
+  //crate orderItemsk
+  async createOrderItem(data: CreateOrderItemDto) {
+    const orderItem = await this.prisma.orderItem.create({
+      data,
+    });
     return orderItem;
   }
   async getOrderItemById(id: string) {
-    const OrderItem = await this.prisma.orderItem.findFirst({
+    if (isNaN(parseInt(id)))
+      throw new BadRequestException('ID format not found');
+    const OrderItem = await this.prisma.orderItem.findUnique({
       where: {
         id: parseInt(id),
       },
     });
+    if (!OrderItem) throw new NotFoundException('OrderItem not found');
     return OrderItem;
   }
   async getOrderItems() {
@@ -22,6 +32,13 @@ export class OrderitemService {
     return orderItems;
   }
   async deleteOrderItemById(id: string) {
+    if (isNaN(parseInt(id))) throw new BadRequestException('Invalid ID format');
+    const orderItemCheck = await this.prisma.orderItem.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+    if (!orderItemCheck) throw new NotFoundException('orderItem not found ');
     const orderitem = await this.prisma.orderItem.delete({
       where: {
         id: parseInt(id),
