+++ backend/src/user/user.service.ts
@@ -1,27 +1,43 @@
-import { Injectable } from '@nestjs/common';
-import { Prisma } from '@prisma/client';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { createUserDto } from './dto/create-user.dto';
+import { CreateUserDto } from './dto/create-user.dto';
+import { User } from '@prisma/client';
+import bcrypt from 'bcrypt';
+import { NotFoundError } from 'rxjs';
 
 @Injectable()
 export class UserService {
   constructor(private prisma: PrismaService) {}
-  async createUser(data: createUserDto) {
-    const user = await this.prisma.user.create({ data });
-    console.log(user);
-    return 'hey';
+
+  async createUser(data: CreateUserDto): Promise<Omit<User, 'password'>> {
+    const hashedPassword = await bcrypt.hash(data.password, 10);
+    const user = await this.prisma.user.create({
+      data: {
+        ...data,
+        password: hashedPassword,
+      },
+    });
+    const { password, ...result } = user;
+    return result;
   }
-  async getusers() {
+  async getUsers(): Promise<User[] | null> {
     const users = await this.prisma.user.findMany();
     return users;
   }
-  async getUserById(id: string) {
-    const users = await this.prisma.user.findUnique({
+  async getUserById(id: string): Promise<User | null> {
+    const user = await this.prisma.user.findUnique({
       where: { id: parseInt(id, 10) },
     });
-    console.log(users);
+    console.log(user);
+    return user;
   }
-  async DeleteUser(id: string) {
+  async deleteUser(id: string): Promise<void> {
+    const user = await this.prisma.user.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+    if (!user) throw new NotFoundException('User not found');
     await this.prisma.user.delete({
       where: {
         id: parseInt(id),
